requirements.txt
altair==5.4.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
gitdb==4.0.12
gitpython==3.1.45
idna==3.11
importlib-resources==6.4.5
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
metatrader5==5.0.5388
mplfinance==0.12.10b0
narwhals==1.42.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
plotext==5.3.2
plotly==6.4.0
protobuf==5.29.5
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.35.1
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
zipp==3.20.2
